Undo renaming io.py
